Fix for bug #51847 Someone in libintl world decided it was a good idea to do:
#undef setlocale
#define setlocale libintl_setlocale

That means, of course, that any use of "setlocale" by the
preprocessor now becomes libintl_setlocale which means we
no longer have a _zif_setlocale symbol in string.o, but we
magically have a _zif_libintl_setlocale symbol.  The linker
is obviously not very happy about this when basic_functions.o
tries to find _zif_setlocale.

So, let's not do that again.



git-svn-id: http://svn.php.net/repository/php/php-src/php/php-src/branches/PHP_5_2@300085 c90b9560-bf6c-de11-be94-00142212c4b1